refactor bases to get working on the gpu

+++ test/unit_tests/UnitTestUtils.cpp
@@ -332,7 +332,7 @@ void CheckQuadraticCompleteness(const Eigen::Matrix<double, Eigen::Dynamic, 1>& 
     for (size_t i = 0; i < 3; ++i) {
         for (size_t j = 0; j < 3; ++j) {
             double calculated_value = 0.0;
-            for (size_t k = 0; k < shape_functions.rows(); ++k) {
+            for (size_t k = 0; k < static_cast<size_t>(shape_functions.rows()); ++k) {
                 calculated_value += shape_functions(k) * neighbor_coordinates(k, i) * neighbor_coordinates(k, j);
             }
             EXPECT_NEAR(calculated_value, evaluation_point_phyiscal_coordinates(0, i) * evaluation_point_phyiscal_coordinates(0, j), tolerance);
@@ -348,7 +348,7 @@ void CheckCubicCompleteness(const Eigen::Matrix<double, Eigen::Dynamic, 1>& shap
         for (size_t j = 0; j < 3; ++j) {
             for (size_t k = 0; k < 3; ++k) {
                 double calculated_value = 0.0;
-                for (size_t l = 0; l < shape_functions.rows(); ++l) {
+                for (size_t l = 0; l < static_cast<size_t>(shape_functions.rows()); ++l) {
                     calculated_value += shape_functions(l) * neighbor_coordinates(l, i) * neighbor_coordinates(l, j) * neighbor_coordinates(l, k);
                 }
                 EXPECT_NEAR(calculated_value, evaluation_point_phyiscal_coordinates(0, i) * evaluation_point_phyiscal_coordinates(0, j) * evaluation_point_phyiscal_coordinates(0, k), tolerance);
@@ -366,7 +366,7 @@ void CheckQuarticCompleteness(const Eigen::Matrix<double, Eigen::Dynamic, 1>& sh
             for (size_t k = 0; k < 3; ++k) {
                 for (size_t l = 0; l < 3; ++l) {
                     double calculated_value = 0.0;
-                    for (size_t m = 0; m < shape_functions.rows(); ++m) {
+                    for (size_t m = 0; m < static_cast<size_t>(shape_functions.rows()); ++m) {
                         calculated_value += shape_functions(m) * neighbor_coordinates(m, i) * neighbor_coordinates(m, j) * neighbor_coordinates(m, k) * neighbor_coordinates(m, l);
                     }
                     EXPECT_NEAR(calculated_value, evaluation_point_phyiscal_coordinates(0, i) * evaluation_point_phyiscal_coordinates(0, j) * evaluation_point_phyiscal_coordinates(0, k) * evaluation_point_phyiscal_coordinates(0, l), tolerance);
